requirements.txt
Django>=5.2.6
blockfrost-python>=0.6.0
pycardano>=0.10.0
python-decouple>=3.8
drf-spectacular>=0.26.0
